Update README.md
Introduction to ATM Program in Java
We can construct an ATM program in Java to display ATM transactions. An automated teller machine (ATM) or cash machine (In British English) is an electronic telecommunications system that allows customers of banking firms to conduct financial transactions. The user must choose a choice from the options shown on the screen in the ATM application. For example, withdraw money, deposit money, check your balance, and exit the available options.

Working of ATM Program in Java
To withdraw your funds, deposit your funds and check your account balance before exiting, the following operations are needed to perform in the ATM program:

Withdraw: For withdrawing the funds, gets the withdrawal amount from the user, deduct it from the total balance, and display the message.
Deposit: For depositing the funds, gets the deposit amount from the user to add, add it to the total balance, and display the message.
Check the balance: For checking the balance, display the user’s total balance.
Exit: Return the user to the home page or initial screen by exiting the current Transaction mode.